requirements.txt
Flask==0.10.1
Flask-Failsafe==0.2
geojson==1.3.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.1
PyYAML==3.11
Shapely==1.5.13
Werkzeug==0.11.4
geojson
psycopg2
shapely
pyyaml
